wwassignment4 is a version of "bridge1" which is compatible with moodle 1.9 and webwork 2.4.x
It adds a new assignment like activity called wwassignment.

It works but does not yet have much testing and has the potential that
the webwork and moodle databases might get out of sync.

The structure is similar enough that it might eventually be a good idea
to implement it as a new assignment type.

Bridge2 adds a question type to the moodle quiz.  It is requires a different
set of software


git-svn-id: http://svn.webwork.maa.org/system/trunk/wwmoodle@5674 c0722133-6baf-4dd8-8699-98d999cd4f06